+++ croud/clusters/commands.py
@@ -442,6 +442,54 @@ def clusters_set_backup_schedule(args: Namespace) -> None:
     )
 
 
+def clusters_snapshots_list(args: Namespace) -> None:
+    url = f"/api/v2/clusters/{args.cluster_id}/snapshots/"
+
+    client = Client.from_args(args)
+    data, errors = client.get(url)
+    print_response(
+        data=data,
+        errors=errors,
+        keys=[
+            "created",
+            "repository",
+            "snapshot",
+            "tables",
+            "concrete_indices",
+        ],
+        output_fmt=get_output_format(args),
+    )
+
+
+def clusters_snapshots_restore(args: Namespace) -> None:
+    url = f"/api/v2/clusters/{args.cluster_id}/snapshots/restore/"
+
+    body = {
+        "repository": args.repository,
+        "snapshot": args.snapshot,
+    }
+    if args.source_cluster_id:
+        body["source_cluster_id"] = args.source_cluster_id
+    if args.tables:
+        tables = args.tables.strip().split(",")
+        body["tables"] = [x.strip() for x in tables]
+
+    client = Client.from_args(args)
+    data, errors = client.post(url, body=body)
+    print_response(
+        data=data,
+        errors=errors,
+        keys=[
+            "created",
+            "repository",
+            "snapshot",
+            "tables",
+            "concrete_indices",
+        ],
+        output_fmt=get_output_format(args),
+    )
+
+
 # We want to map the custom hardware specs to slightly nicer params in croud,
 # hence this mapping here
 def _handle_edge_params(body, args):